+++ app/views/users/_user_referents_list.html.erb
@@ -11,7 +11,7 @@
           </span>
           <%=
             link_to(
-             structure_referent_assignations_path(user.id, agent_id: referent.id),
+             structure_referent_assignation_path(user.id, referent.id),
               data: {
                 turbo_confirm: "Retirer l'agent référent de cet usager ?",
                 turbo_confirm_text_content: "Êtes-vous sûr de vouloir retirer l'agent référent <b>#{referent}</b> de l'usager <b>#{user}</b> ?",
