+++ app/views/users_tags/_user_tag_badge.html.erb
@@ -10,7 +10,7 @@
   <% if with_delete %>
     <%=
       link_to(
-        structure_tag_assignations_path(user.id, tag_id: tag.id),
+        structure_tag_assignation_path(user.id, tag.id),
         data: {
           turbo_confirm: "Retirer le tag de cet usager ?",
           turbo_confirm_text_content: "Êtes-vous sûr de vouloir retirer le tag <b>#{tag.value}</b> de l'usager <b>#{user}</b> ?",
