+++ config/routes.rb
@@ -55,8 +55,6 @@ Rails.application.routes.draw do
       scope module: :users do
         resources :follow_ups, only: [:index]
         resource :parcours, only: [:show]
-        resources :orientations, only: [:new, :create, :edit, :update, :destroy]
-        resources :parcours_documents, only: [:show, :update, :create, :destroy]
       end
       resources :invitations, only: [:create]
     end
@@ -67,12 +65,7 @@ Rails.application.routes.draw do
     resources :category_configurations, only: [:index, :show, :new, :create, :edit, :update, :destroy]
     patch "category_configurations_positions/update", to: "category_configurations_positions#update"
     resources :tags, only: [:create, :destroy]
-    resources :users_organisations, only: [:index, :create]
-    resource :users_organisations, only: [:destroy]
-    resources :referent_assignations, only: [:index, :create]
-    resource :referent_assignations, only: [:destroy]
-    resources :tag_assignations, only: [:index, :create]
-    resource :tag_assignations, only: [:destroy]
+    resources :users_organisations, only: [:index, :create, :destroy]
     resources :agent_roles, module: :agent_roles, only: [] do
       collection do
         resources :csv_export_authorizations, only: [:index]
@@ -93,6 +86,10 @@ Rails.application.routes.draw do
 
   resources :users, module: :users, only: [] do
     resources :rdvs, only: [:new]
+    resources :parcours_documents, only: [:show, :update, :create, :destroy]
+    resources :tag_assignations, only: [:index, :create, :destroy], param: :tag_id
+    resources :referent_assignations, only: [:index, :create, :destroy], param: :agent_id
+    resources :orientations, only: [:new, :create, :edit, :update, :destroy]
   end
 
   get "invitation", to: "invitations#invitation_code", as: :invitation_landing
@@ -146,8 +143,6 @@ Rails.application.routes.draw do
       scope module: :users do
         resources :follow_ups, only: [:index]
         resource :parcours, only: [:show]
-        resources :orientations, only: [:new, :create, :edit, :update, :destroy]
-        resources :parcours_documents, only: [:show, :update, :create, :destroy]
       end
       resources :invitations, only: [:create]
       resources :archives, only: [:new, :create]
@@ -158,10 +153,6 @@ Rails.application.routes.draw do
     resource :stats, only: [:show], controller: 'website/stats'
     resources :users_organisations, only: [:index, :create]
     resource :users_organisations, only: [:destroy]
-    resources :referent_assignations, only: [:index, :create]
-    resource :referent_assignations, only: [:destroy]
-    resources :tag_assignations, only: [:index, :create]
-    resource :tag_assignations, only: [:destroy]
     resources :invitation_dates_filterings, :creation_dates_filterings, only: [:new]
   end
